requirements.txt
langchain
langchain-google-genai
langchain-community
python-dotenv
